services/api/pyproject.toml
[project]
name = "heimdex-api"
version = "0.1.0"
description = "Heimdex API service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "psycopg[binary]>=3.1.0",
    "supabase>=2.3.0",
    "dramatiq[redis]>=1.15.0",
    "redis>=5.0.0",
    "openai>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.26.0",
]
